requirements.txt
python-docx
docx2pdf==0.1.8
elevenlabs==1.2.1
faiss-cpu
ffmpeg_python==0.2.0
ipython==8.17.2
langchain==0.1.16
langchain_community==0.0.34
langchain_core==0.1.46
langchain_openai==0.1.4
librosa==0.9.2
moviepy==1.0.3
mutagen==1.47.0
openai
pandas==2.0.3
# Pillow==10.0.1
Pillow==10.3.0
pydub==0.25.1
python-dotenv==1.0.1
pytube==15.0.0
pytube3==9.6.4
SpeechRecognition==3.10.0
streamlit==1.33.0
streamlit-chat==0.1.1
streamlit-file-browser==3.2.22
streamlit-pdf-viewer==0.0.10
streamlit-pills==0.3.0
tiktoken
tqdm==4.66.1
youtube-search-python==1.6.6
youtube-transcript-api==0.6.1
yt-dlp==2023.10.13
fpdf2
PyPDF2<3.0.0